made the public functions in their own files
