+++ back_end/models/item.py
@@ -53,25 +53,3 @@ class Item(ItemBase, table=True):
     title: str = Field(max_length=255)
     owner_id: int | None = Field(default=None, foreign_key="user.id", nullable=False)
     owner: User | None = Relationship(back_populates="items")
-
-
-class ItemPublic(ItemBase):
-    """Properties to return via API for public item information.
-
-    Attributes:
-        id (int): The unique identifier for the item.
-    """
-
-    id: int
-
-
-class ItemsPublic(SQLModel):
-    """Properties to return via API for a list of public items.
-
-    Attributes:
-        data (list[ItemPublic]): The list of public item information.
-        count (int): The total number of items.
-    """
-
-    data: list[ItemPublic]
-    count: int
