+++ back_end/models/user.py
@@ -109,23 +109,3 @@ class User(UserBase, table=True):
 
     def hash_password(self):
         self.hashed_password = get_password_hash(self.password)
-
-
-class UserPublic(UserBase):
-    """Properties to return via API for public user information.
-
-    Attributes:
-        id (int): The unique identifier for the user.
-    """
-
-    id: int
-
-
-class UsersPublic(SQLModel):
-    """Properties to return via API for a list of public users.
-
-    Attributes:
-        users (list[UserPublic]): The list of public user information.
-    """
-
-    users: list[UserPublic]
